Transcripcion/speech2-main/requirements.txt
functions-framework
google-cloud-speech
google-cloud-storage
joblib
nltk
scikit-learn
pandas
numpy
dill
lime

